Fix -Wshorten-64-to-32 on 64bit backend code part 1
This CL fixes backend code (except d3d backends).

Bug: 523358231
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/328496
Commit-Queue: Shrek Shao <shrekshao@google.com>

+++ src/dawn/native/metal/BufferMTL.mm
@@ -140,9 +140,9 @@ MaybeError Buffer::Initialize(bool mappedAtCreation) {
 
     // Initialize the padding bytes to zero.
     if (GetDevice()->IsToggleEnabled(Toggle::LazyClearResourceOnFirstUse) && !mappedAtCreation) {
-        uint32_t paddingBytes = GetAllocatedSize() - GetSize();
+        size_t paddingBytes = GetAllocatedSize() - GetSize();
         if (paddingBytes > 0) {
-            uint32_t clearSize = Align(paddingBytes, 4);
+            size_t clearSize = Align(paddingBytes, 4);
             uint64_t clearOffset = GetAllocatedSize() - clearSize;
 
             auto scopedUseDuringCreation = UseInternal();
