+++ src/dawn/native/opengl/ShaderModuleGL.cpp
@@ -47,6 +47,7 @@
 #include "src/dawn/native/opengl/PipelineLayoutGL.h"
 #include "src/dawn/native/opengl/UtilsGL.h"
 #include "src/dawn/platform/tracing/TraceEvent.h"
+#include "src/utils/numeric.h"
 #include "tint/tint.h"
 
 namespace dawn::native::opengl {
@@ -452,7 +453,8 @@ ResultOrError<GLuint> ShaderModule::CompileShader(
     // element byte size and pass that to Tint.
     auto immediateCount = RoundUp(pipelineImmediateMask.count(), 4u);
 
-    req.tintOptions.minimum_immediate_size = immediateCount * kImmediateElementByteSize;
+    req.tintOptions.minimum_immediate_size =
+        checked_cast<uint32_t>(immediateCount * kImmediateElementByteSize);
     if (HasImmediates(&RenderImmediates::firstVertex, pipelineImmediateMask)) {
         req.tintOptions.first_vertex_offset = GetImmediateByteOffsetInPipelineIfAny(
             &RenderImmediates::firstVertex, pipelineImmediateMask);
