+++ src/dawn/native/vulkan/ShaderModuleVk.cpp
@@ -62,6 +62,7 @@
 #include "src/dawn/native/vulkan/VulkanError.h"
 #include "src/dawn/platform/metrics/HistogramMacros.h"
 #include "src/dawn/platform/tracing/TraceEvent.h"
+#include "src/utils/numeric.h"
 #include "tint/tint.h"
 
 #ifdef DAWN_ENABLE_SPIRV_VALIDATION
@@ -367,7 +368,8 @@ ResultOrError<ShaderModule::ModuleAndSpirv> ShaderModule::GetHandleAndSpirv(
     // Set internal immediate offsets
     // Size the immediate block to the pipeline's used slots so the decomposed array matches the
     // push constant range reserved by the pipeline layout (see ToPushConstantBytes).
-    req.tintOptions.minimum_immediate_size = in.immediateMask.count() * kImmediateElementByteSize;
+    req.tintOptions.minimum_immediate_size =
+        checked_cast<uint32_t>(in.immediateMask.count() * kImmediateElementByteSize);
     if (HasImmediates(&RenderImmediates::clampFragDepth, in.immediateMask)) {
         uint32_t offsetStartBytes =
             GetImmediateByteOffsetInPipeline(&RenderImmediates::clampFragDepth, in.immediateMask);
